+++ internal/es/execution/execution_in_memory.go
@@ -4,12 +4,11 @@ import (
 	"context"
 	"encoding/json"
 	"fmt"
-	"github.com/turbot/pipe-fittings/sanitize"
-	"io"
 	"log/slog"
-	"os"
 	"strings"
 	"time"
+
+	"github.com/turbot/pipe-fittings/sanitize"
 
 	"github.com/turbot/go-kit/helpers"
 
@@ -18,7 +17,7 @@ import (
 	"github.com/turbot/flowpipe/internal/cache"
 	"github.com/turbot/flowpipe/internal/es/db"
 	"github.com/turbot/flowpipe/internal/es/event"
-	"github.com/turbot/flowpipe/internal/filepaths"
+	"github.com/turbot/flowpipe/internal/store"
 	pfconstants "github.com/turbot/pipe-fittings/constants"
 	"github.com/turbot/pipe-fittings/funcs"
 	"github.com/turbot/pipe-fittings/hclhelpers"
@@ -86,43 +85,10 @@ func GetPipelineDefnFromExecution(executionID, pipelineExecutionID string) (*Exe
 	return ex, defn, nil
 }
 
-func (ex *ExecutionInMemory) SaveToFile() error {
-	eventStoreFilePath := filepaths.EventStoreFilePath(ex.ID)
-
-	stat, _ := os.Stat(eventStoreFilePath)
-	if stat != nil {
-		// Keeping this simple, we don't want to overwrite the file. This may change in the future when we can resume execution.
-		// Right now if Flowpipe stops/crashes there's no way to resume the execution because it's no longer in memory and not
-		// persisted
-		return perr.BadRequestWithMessage("execution file already exists. execution can only be serialised once at termination")
-	}
-
-	// Append the JSON data to a file
-	file, err := os.OpenFile(eventStoreFilePath, os.O_APPEND|os.O_CREATE|os.O_WRONLY, 0644)
+func (ex *ExecutionInMemory) Save() error {
+	err := ex.saveToSQLite()
 	if err != nil {
-		return perr.InternalWithMessage("Error opening file " + err.Error())
-	}
-	defer file.Close()
-
-	for _, event := range ex.Events {
-		// Marshal the struct to JSON
-		eventData, err := json.Marshal(event) // No indent, single line
-		if err != nil {
-			slog.Error("Error marshalling JSON", "error", err)
-			return err
-		}
-
-		sanitizedEventData := sanitize.Instance.SanitizeString(string(eventData))
-
-		_, err = file.Write([]byte(sanitizedEventData))
-		if err != nil {
-			return perr.InternalWithMessage("Error writing to file " + err.Error())
-		}
-
-		_, err = file.WriteString("\n")
-		if err != nil {
-			return perr.InternalWithMessage("Error writing to file " + err.Error())
-		}
+		return err
 	}
 
 	// This seems a convenient place to expire the execution from the cache
@@ -130,6 +96,45 @@ func (ex *ExecutionInMemory) SaveToFile() error {
 	if err != nil {
 		slog.Error("Error completing execution", "error", err)
 		return err
+	}
+
+	return nil
+}
+
+func (ex *ExecutionInMemory) saveToSQLite() error {
+	db, err := store.OpenFlowpipeDB()
+	if err != nil {
+		return perr.InternalWithMessage("Error opening SQLite database " + err.Error())
+	}
+	defer db.Close()
+
+	tx, err := db.Begin()
+	if err != nil {
+		return perr.InternalWithMessage("Error beginning SQLite transaction " + err.Error())
+	}
+
+	for _, event := range ex.Events {
+
+		// Marshall the payloa to JSON
+
+		payloadData, err := json.Marshal(event.Payload)
+		if err != nil {
+			slog.Error("Error marshalling JSON", "error", err)
+			return err
+		}
+
+		sanitizePayloadData := sanitize.Instance.SanitizeString(string(payloadData))
+
+		statement := `INSERT INTO event (execution_id, created_at, type, data) values (?, ?, ?, ?)`
+		_, err = db.Exec(statement, ex.ID, event.Timestamp, event.EventType, sanitizePayloadData)
+		if err != nil {
+			return err
+		}
+	}
+
+	err = tx.Commit()
+	if err != nil {
+		return perr.InternalWithMessage("Error committing SQLite transaction " + err.Error())
 	}
 
 	return nil
@@ -466,25 +471,6 @@ func (ex *ExecutionInMemory) ProcessEvents() error {
 	}
 	ex.LastProcessedEventIndex = len(ex.Events)
 
-	return nil
-}
-
-// LoadFromFile loads an execution from a JSON file.
-func (ex *ExecutionInMemory) LoadJSON(fileName string) error {
-	jsonFile, err := os.Open(fileName)
-	if err != nil {
-		return err
-	}
-	// defer the closing of our jsonFile so that we can parse it later on
-	defer jsonFile.Close()
-	byteValue, err := io.ReadAll(jsonFile)
-	if err != nil {
-		return err
-	}
-	err = json.Unmarshal(byteValue, &ex)
-	if err != nil {
-		return err
-	}
 	return nil
 }