+++ internal/es/handler/pipeline_failed.go
@@ -77,7 +77,7 @@ func (h PipelineFailed) Handle(ctx context.Context, ei interface{}) error {
 		output.RenderServerOutput(ctx, o)
 	}
 
-	err = ex.SaveToFile()
+	err = ex.Save()
 	if err != nil {
 		slog.Error("pipeline_finished: Error saving execution", "error", err)
 		// Should we raise pipeline fail here?